feat: cleanup of the repository
This is initial cleanup of the Coldfire repository. Functions are spread between files in logical manner.
In the future, instead of files lied flat in coldfire package, subpackages should be created.

Some basic tests are added, additionally created GitHub Actions workflows.

Closes redcode-labs/coldfire#27
