+++ .github/workflows/test.yml
@@ -1,0 +1,25 @@
+on:
+  pull_request:
+    branches:
+      - master
+name: Test
+
+jobs:
+
+  test:
+    runs-on: ubuntu-latest
+    steps:
+    - name: Install Go
+      uses: actions/setup-go@v2
+      with:
+        go-version: 1.17.x
+    - name: Checkout code
+      uses: actions/checkout@v2
+    - name: Test
+      run: |
+        go test -coverprofile coverage.txt -covermode atomic ./...
+    - name: Upload report
+      env:
+        CODECOV_TOKEN: "{{ secrets.CODECOV_TOKEN }}"
+      run: |
+        bash <(curl -s https://codecov.io/bash)
