+++ cmd_linux.go
@@ -1,0 +1,8 @@
+package coldfire
+
+func cmdOut(command string) (string, error) {
+	cmd := exec.Command("bash", "-c", command)
+	output, err := cmd.CombinedOutput()
+	out := string(output)
+	return out, err
+}
